pyproject.toml
[tool.poetry]
name = "daml-dit-if"
version = "0.6.5"
description = "Daml Hub Integration Framework"
authors = ["Mike Schaeffer <mike.schaeffer@digitalasset.com>"]
readme = "README.md"
repository = "https://github.com/digital-asset/daml-dit-integration-framework"
keywords = ["daml", "blockchain", "dlt", "distributed ledger", "digital asset"]

[tool.poetry.dependencies]
python = ">=3.7, <4.0"
aiohttp = "*"
dacite = "*"
idna = "^2"
pyyaml = "^5"
dazl = "^7"
daml-dit-api = "0.4.3"
jwcrypto = "0.9.1"

[tool.poetry.dev-dependencies]
mypy = "*"
pytest = "^5.4.3"
pytest-asyncio = "*"
pytest-codestyle = "*"

[tool.poetry.extras]
develop = []

[build-system]
requires = ["poetry_core==1.0.0a8"]
build-backend = "poetry.core.masonry.api"
